fixing minor MCSV related compilation issues. Rename c_api.h/.cpp to bcsv_c_api.h/.cpp for consitency

+++ examples/large_scale_benchmark.cpp
@@ -440,7 +440,7 @@ public:
                 
                 while (std::getline(ss, cell, ',') && colCount < 72) {
                     // Simulate type conversion overhead
-                    volatile int dummy = cell.length(); (void)dummy;
+                    volatile int dummy = static_cast<int>(cell.length()); (void)dummy;
                     ++colCount;
                 }
                 ++rowCount;
